+++ doc/change-log.md
@@ -1,6 +1,13 @@
+# v2.2.0 [2024-03-05]
+## Improvement
+- 引入 `com.lucendar.common.types.Validatable`
+
+## Fixed
+- 将 `LinesValidateResultReceiver.invalidField` 抛出的异常代码修正为 INVALID_CONFIG，而非 INVALID_PARAM
+
 # v2.1.0
 ## Improvement
-- Add `Errors.OBJECT_NOT_FOUND`
+- 新增 `Errors.OBJECT_NOT_FOUND`
 
 # v2.0.0
 ## Improvement